More robust error handling and informing user of errors - Errors in a drawable can now trickle up all the way to the InteractiveApplication - They are not thrown (so as not to break other apps) but are instead stored and can be retrieved via accessor functions - Presently, the InteractiveApplication run loop will check for an errror in the current drawable through the renderable - If one is detected it clears it and then shows it to the user via a QMessageBox
WARNING: tetzlaff.interactive.InteractiveApplication now uses QMessageBox and is therefore dependent on QtJambi.  This
may be undesirable but I could not find a better way to be able to inform the user of an error without deeper changes.
Perhaps a callback approach could be used where the top level application could define a method for displaying a message
box that uses the API of their choice.  This is notated in the code.