Refactored specular fit to allow a complete solution to be loaded when a project is loaded.
Hasn't been tested yet.
TODO:
Force user to save project as soon as the view set file is loaded, otherwise textures will be saved to the Metashape project directory.  At that time the texture fit directory also needs to be set.
The whole "root directory" system for view sets may need to be overhauled.
Saving textures should also really be made more modular so that we can easily save everything do a new directory when the user does "Save as" (and not save by default if they have autosave off).