Created the FXML file
Created the FXML file for the export request and formatted it how it was previously.
In the java file on start up it sets all the values to how the settings has it, as well as when you run the program it sets the settings to what the user requested.