Final push for my hours
Pushing so I don't break it before I try more with the lambda expression.

Style sheet has all the changes that I have gotten over the time. Made the menu buttons square and a few more changes.
Eye dropper is how it looks like on the figma, except for the images. If I switch the rectangles to images they lose the functionality and i don't have time to rewrite it.
menu bar has eyedropper in scene now, but it is invisible right away, and if you open it there is no way to get rid of it currently.